% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.plot.R
\name{dot.plot}
\alias{dot.plot}
\title{Plot variable means across time}
\usage{
dot.plot(data, params, species = "all", trendline = TRUE,
  save2disk = FALSE, dir.out = NULL, wcm = 27, hcm = 19)
}
\arguments{
\item{data}{The data.frame with the trapping data}

\item{params}{The variables to be plotted}

\item{species}{The species to be included in the plots. If "all", then all
the species found will be plotted}

\item{trendline}{Whether to plot a trend line (a loeass curve)}

\item{save2disk}{Whether to save files to disk (default FALSE)}

\item{dir.out}{The path where to save the files if save2disk=TRUE}

\item{wcm}{width of saved plots in cm}

\item{hcm}{height of saved plots in cm}
}
\value{
A list with a plot for each parameter. The last element is a table
with mean and confidence intervals for each parameter
}
\description{
\code{dot.plot} plots the means of the variables passed with \code{params}.
1.96 x SE (confidence intervals) bars are also plotted.
}
\details{
If \code{save2disk==TRUE} a PDF and RDA file will be saved to disk in the
directory passed with \code{dir.out} (which will be created if doesn't exist).
Size of the saved plots can be controlled with \code{wcm} and \code{hcm}.
Default settings save plots with a landscape orientation on A4 size.
}
