% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTBA.R
\name{KTBA}
\alias{KTBA}
\title{A function to compute the number of animal known to be alive and plot it over
time for each species}
\usage{
KTBA(data, species = "all", animal.ids = "M.chip.serial..")
}
\arguments{
\item{data}{The data.frame with the trapping data}

\item{species}{The species to be included in the plots. If "all", then all
the species found will be plotted}

\item{animal.ids}{the header of the column with the animals' IDs}
}
\value{
A list wher the first element is a plot and the second are the data
}
\description{
A function to compute the number of animal known to be alive and plot it over
time for each species
}
