% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mark.inp.R
\name{make.mark.inp}
\alias{make.mark.inp}
\title{Make a Mark input file}
\usage{
make.mark.inp(fn, h = FALSE, ids = TRUE, count = 1)
}
\arguments{
\item{fn}{File name}

\item{h}{whether the .csv has headers}

\item{ids}{whether the first column is animals' ids}

\item{count}{the count of animals with each history}
}
\value{
saves a .inp file and returns a character vector with the capture
  history
}
\description{
Save a text file with extension .inp with the capture history formatted as
input for the program Mark from data with capture history with a column for
each capture session
}
