% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.dist.plot.R
\name{param.dist.plot}
\alias{param.dist.plot}
\title{Plot variable distribution}
\usage{
param.dist.plot(data, params, species = "all", save2disk = FALSE,
  dir.out = NULL)
}
\arguments{
\item{data}{The data.frame with the trapping data}

\item{params}{The variables to be plotted}

\item{species}{The species to be included in the plots. If "all", then all
the species found will be plotted}

\item{save2disk}{Whether to save files to disk (default FALSE)}

\item{dir.out}{The path where to save the files if save2disk=TRUE}
}
\value{
A list with a plot for each parameter.
}
\description{
\code{param.dist.plot} generates an histogram of the variables passed with
\code{params}. Data are separated by sex and species.
}
\details{
If \code{save2disk==TRUE} a PDF and RDA file will be saved to disk in the
directory passed with \code{dir.out} (which will be created if doesn't exist).
}

