% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.means.R
\name{plot.means}
\alias{plot.means}
\title{Plot variable means across time}
\usage{
\method{plot}{means}(data, params, species = "all", save2disk = FALSE,
  dir.out = NULL)
}
\arguments{
\item{data}{The data.frame with the trapping data}

\item{params}{The variables to be plotted}

\item{species}{The species to be included in the plots. If "all", then all
the species found will be plotted}

\item{save2disk}{Whether to save files to disk (default FALSE)}

\item{dir.out}{The path where to save the files if save2disk=TRUE}
}
\description{
\code{plot.means} plots the means of the variables passed with \code{params}.
1.96 x SE bars are also plotted.
}
\details{
If \code{save2disk==TRUE} a PDF and RDA file will be saved to disk in the
directory passed with \code{dir.out}.
}

