% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mark.inp.R
\name{trap2mark}
\alias{trap2mark}
\title{Transform WP trapping database into mark capture history input file}
\usage{
trap2mark(data, species = "all", ndays = 14,
  animal.ids = "M.chip.serial..", run = "Run")
}
\arguments{
\item{data}{The data.frame with the trapping data}

\item{species}{The species to be included in the plots. If "all", then all
the species found will be plotted}

\item{ndays}{The time difference between two trapping sessions to be
considered independent}

\item{animal.ids}{the column header that identifies the animals' IDs}

\item{run}{the column header that identifies different trapping run within
the same day}
}
\value{
a list with an element for each \code{species} where each element is
  a list with the capture history for each session as elements
}
\description{
Transform WP trapping database into mark capture history input file
}
