% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trap.year.R
\name{trap.year}
\alias{trap.year}
\title{Trapping across years}
\usage{
trap.year(data)
}
\arguments{
\item{data}{The input data (output from \code{read.trapping})}
}
\value{
A data.frame with years for column and species as rows
}
\description{
\code{trap.year} calculates the number of sessions within each year, for each
species.
}

